+++ azure-mgmt-servicebus/src/main/java/com/microsoft/azure/management/servicebus/Subscription.java
@@ -55,11 +55,11 @@ public interface Subscription extends
     /**
      * @return the duration of peek-lock which is the amount of time that the message is locked for other receivers
      */
-    int lockDurationInSeconds();
+    long lockDurationInSeconds();
     /**
      * @return the idle duration after which the subscription is automatically deleted.
      */
-    int deleteOnIdleDurationInMinutes();
+    long deleteOnIdleDurationInMinutes();
     /**
      * @return the duration after which the message expires, starting from when the message is sent to subscription.
      */
@@ -71,28 +71,28 @@ public interface Subscription extends
     /**
      * @return the number of messages in the subscription
      */
-    int messageCount();
+    long messageCount();
     /**
      * @return number of active messages in the subscription
      */
-    int activeMessageCount();
+    long activeMessageCount();
     /**
      * @return number of messages in the dead-letter subscription
      */
-    int deadLetterMessageCount();
+    long deadLetterMessageCount();
     /**
      * @return number of messages sent to the subscription that are yet to be released
      * for consumption
      */
-    int scheduledMessageCount();
+    long scheduledMessageCount();
     /**
      * @return number of messages transferred into dead letters
      */
-    int transferDeadLetterMessageCount();
+    long transferDeadLetterMessageCount();
     /**
      * @return number of messages transferred to another queue, topic, or subscription
      */
-    int transferMessageCount();
+    long transferMessageCount();
     /**
      * @return the current status of the subscription
      */
